pyproject.toml
[project]
name = "kb-real-estate-data"
version = "0.1.0"
description = "KB 부동산 및 공공데이터 수집 도구"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gspread>=6.2.1",
    "gspread-dataframe>=4.0.0",
    "jupyter>=1.1.1",
    "pandas>=2.3.3",
    "publicdatareader>=1.1.0",
    "pygwalker>=0.4.9.15",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
